refactoring name Service, create QuoteRepository, create BaseRepository, create model Quote, create Quotation, create BaseModel
