+++ app/Models/BaseModel.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace App\Models;
+
+class BaseModel
+{
+
+}
