+++ app/Repositories/BaseRepository.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace App\Repositories;
+
+class BaseRepository
+{
+
+}
